+++ semcore/popper/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.24.0] - 2024-03-26
+
+### Fixed
+
+- Focus lock with `disablePortal` prop.
+
 ## [5.23.0] - 2024-03-19
 
 ### Changed