+++ semcore/utils/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.22.0] - 2024-03-26
+
+### Added
+
+- Internal utility for uncontrolled props handling.
+
 ## [4.21.1] - 2024-03-19
 
 ### Fixed